Migrate bookmark to its own component
