+++ client/src/pages/Search/GridView/RestaurantCard/ImageCard.js
@@ -1,8 +1,5 @@
 import defaultRestaurantImg from "../../../../assets/images/default-restaurant-img.png";
-import { useSelector } from "react-redux";
-import { IconButton } from "@mui/material";
-import BookmarkBorderIcon from "@mui/icons-material/BookmarkBorder";
-import BookmarkIcon from "@mui/icons-material/Bookmark";
+import Bookmark from "../../../../components/Bookmark";
 
 const classes = {
   imageCard: {
@@ -10,17 +7,9 @@ const classes = {
     height: "100%",
     objectFit: "cover",
   },
-  bookmark: {
-    position: "absolute",
-    top: 0,
-    right: "-2px",
-    color: "#91D4FA",
-    padding: 0,
-  },
 };
 
 function ImageCard({ src, restaurantId }) {
-  const user = useSelector((state) => state.user);
   return (
     <div style={{ position: "relative" }}>
       <img
@@ -31,16 +20,7 @@ function ImageCard({ src, restaurantId }) {
           e.currentTarget.src = defaultRestaurantImg;
         }}
       />
-
-      {user.saved_restaurants.includes(restaurantId) ? (
-        <IconButton style={classes.bookmark}>
-          <BookmarkIcon />
-        </IconButton>
-      ) : (
-        <IconButton style={classes.bookmark}>
-          <BookmarkBorderIcon />
-        </IconButton>
-      )}
+      <Bookmark restaurantId={restaurantId} />
     </div>
   );
 }
